CB-29428 Create RHEL 9 YARN images

+++ scripts/salt-install.sh
@@ -74,7 +74,9 @@ function update_yum_repos() {
       RHEL_VERSION=${RHEL_VERSION/.0/}
       RHEL_VERSION_MAJOR=${RHEL_VERSION:0:1}
       REPO_FILE=rhel${RHEL_VERSION}_cldr_mirrors.repo
-      rm /etc/yum.repos.d/*.repo -f
+      if [ "${CLOUD_PROVIDER}" != "YARN" ]; then
+        rm /etc/yum.repos.d/*.repo -f
+      fi
       curl https://mirror.infra.cloudera.com/repos/rhel/server/${RHEL_VERSION_MAJOR}/${RHEL_VERSION}/${REPO_FILE} --fail > /etc/yum.repos.d/${REPO_FILE}
     fi
   else
